Mutiple consumer group support
This is required so that with differnt consumer groups version could be run together.
Running together is important to save on connections, particularly redshift connection
in the loader.

This would be used by the Operator to start topics whose mask configuration did not change and topics which were reloaded after mask updupdate and has new consumer group names together in one container sharing connection.