Sync loader handler
Fixes https://github.com/practo/tipoca-stream/issues/148
Does not use mbatch library but performs the batch processing instead of flush based on same parameters as before maxSize, maxWaitSeconds and the schemaID. It uses a little different method as described on article "Efficiently consuming Kafka messages". It combines the sync and mutliBatch approach.

https://medium.com/swlh/how-to-consume-kafka-efficiently-in-golang-264f7fe2155b
https://github.com/sceneryback/kafka-best-practices

Still to benchmark on the performance.