Session timeouts, hearbeat, maxProcessing time
New configurations added as based on the type of processing and the size of the batch these values need configurations for every use case. We needed this to process huge batch of loads in Redshift cluster. When the batch size was huge the timeouts where happening leading to commits not going through. Details on why we are doing this is in this issue https://github.com/practo/tipoca-stream/issues/150#issuecomment-790232991