Shutdown processing when session expires.
At present the whole loader will shutdown if session expires for even one routine. We are doing this so that we get to know of the problem fast enough as we get Crashloop alerts on continuous restarts. Later we can change the session timeouts to restart the ConsumeClaim and track the error rate prometheus

https://github.com/practo/tipoca-stream/issues/150#issuecomment-790232991