LoaderCurrentOffset moves out of Group
This was done as the Group in status only stores the information of released stuff. So the realtime calculation needs a separate variable to work on.
Also paves the way to to fix #157. As now the list of DeadConsumerGroups are present in the rsk status. The operator just needs to call Kafka and delete them and clear the list. Cleanup is going to come in separate PR. This helps cleanup as we can safely delete the consumer group as only one topic is only part of one consumer group.