STL scan value is a Gauge
It is Gauge since AWS only keeps last N days data for it and deletes the old. So the queries can go up and down.
We can find a table is in use by just analyzing last X days of data and then use the latest value.

So we make the view have only X days of data and the latest value tells how many times it was accessed.