Fixes loader erroring on deleting distkey column
Redshift by default uses diststyle AUTO and after the last redshift
maintenance update, it seems to make kafkaoffset column as the distkey
which causes the drop column step to fail. This change ensures that for
all tables which don't have an explicit distkey set, it will default to
'diststyle even' and not 'diststyle AUTO'.