Fixes for AQUA breaking scan_query_total view
When AQUA is enabled, the existing view query returns duplicates in the
output due to perm_table_name showing as 'AQUA Scan <table-name>' for
some rows in stl_scan and '<table-name>' for other rows.

This commit fixes the issue by not grouping on perm_table_name and
aggregating just based on table id.